edit product id type

+++ src/main/java/com/capitanbeto/sushi/product/Product.java
@@ -7,7 +7,7 @@ import jakarta.persistence.*;
 public class Product {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(unique = true)
     private String name;
@@ -23,7 +23,7 @@ public class Product {
     public Product() {
     }
 
-    public Product(int id, String name, float price, String category, String description, String image) {
+    public Product(Long id, String name, float price, String category, String description, String image) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -40,11 +40,11 @@ public class Product {
         this.image = image;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
